examples/python/requirements.txt
protobuf==5.29.2
grpcio-tools==1.68.1